+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 11.4.1
 - Fix SQL error invalid input syntax for type integer: "" in Courses.php
 - InputDivOnclick() Do not not convert single quotes to gain a few bytes in Inputs.php
 - Add `functions/PDF.php|pdf_stop_html` & `functions/PDF.php|pdf_stop_pdf` action hooks in Actions.php & PDF.php
+- Fix PHP warning use reset() instead of guessing if $errors[1] is set in ErrorMessage.fnc.php
 
 Changes in 11.4
 ---------------